Move amendments derivation functions
We'd like to add the amendment derivation as a property of NoticeXML; to get
there, we need to disentangle some dependencies, moving those functions from
regparser.notice.build into regparser.notice.amendments
